django_test/requirements.txt
# 환경구성
Django==4.2.3
asgiref==3.7.2
pytz==2023.3
sqlparse==0.4.4
djangorestframework==3.14
django-environ==0.10.0
gunicorn==21.2.0

# db
psycopg2==2.9.6

# 인증
PyJWT==2.8.0
djangorestframework-simplejwt==5.2.2

# test
pytest==7.4.0
pytest-django==4.5.2

# formatter
black==23.7.0
flake8==6.1.0
